backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

langchain-openai
langgraph
langchain
python-docx==1.1.2
pypdf==4.3.1

ddgs
sse-starlette
python-dotenv

# Мониторинг API запросов
mitmproxy
requests